Fix cable event hander bug
Cable event handlers were deleted when the garbade collector deleted the
power grid, and then the cable. Thus the Unity GameObject would be
deleted, and the cable's sprite renderer, but the Cable could still
recieve events from other cables being added / deleted and would try to
render itself.

To fix this, the OnDestroy function was implemented and the event
handlers are now removed there.